+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - `Innmind\BlackBox\Runner\Proof` is now a final class
 - `Innmind\BlackBox\Runner\Proof` constructors are now flagged as internal
 - `Innmind\BlackBox\Runner\Proof::scenarii()` is now flagged as internal
+- PHPUnit attributes `Group` and `DataProvider` are now longer read when running tests with BlackBox, use the `Innmind\BlackBox\` prefixed ones instead
 
 ### Removed